+++ src/containerapp/azext_containerapp/_constants.py
@@ -27,7 +27,7 @@ CONTAINER_APP_EXTENSION_TYPE = "microsoft.app.environment"
 MANAGED_CERTIFICATE_RT = "managedCertificates"
 PRIVATE_CERTIFICATE_RT = "certificates"
 
-DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb"]
+DEV_SERVICE_LIST = ["kafka", "postgres", "redis", "mariadb", "qdrant"]
 
 DEV_KAFKA_IMAGE = 'kafka'
 DEV_KAFKA_SERVICE_TYPE = 'kafka'
@@ -44,6 +44,10 @@ DEV_REDIS_CONTAINER_NAME = 'redis'
 DEV_MARIADB_IMAGE = 'mariadb'
 DEV_MARIADB_SERVICE_TYPE = 'mariadb'
 DEV_MARIADB_CONTAINER_NAME = 'mariadb'
+
+DEV_QDRANT_IMAGE = 'qdrant'
+DEV_QDRANT_SERVICE_TYPE = 'qdrant'
+DEV_QDRANT_CONTAINER_NAME = 'qdrant'
 
 PENDING_STATUS = "Pending"
 SUCCEEDED_STATUS = "Succeeded"