+++ src/containerapp/HISTORY.rst
@@ -4,7 +4,10 @@ Release History
 ===============
 upcoming
 ++++++
+* 'az containerapp service': add support for creation and deletion of Qdrant vector database as a container app dev service
 * Add command group 'az containerapp connected-env', support show/list/delete/create connected environment
+* 'az containerapp create': support --source and --repo properties
+* 'az containerapp update': support --source property
 * 'az containerapp env': --infrastructure-resource-group, supports custom rg name for byovnet env creations in WP enabled envs
 
 0.3.39